[Bench][Timing] fold the CUPTI collector in and cut the tests to what can fail silently
native_cupti.py existed to isolate a compiled extension. That extension is gone,
so the module was a 125-line adapter with one caller and a name that no longer
described it. Folded into benchmark_base.py; NativeCUPTIError and
_NativeCUPTIAttributionError lose the fossil prefix.

test_benchmark_base.py drops to 205 lines. Gone: two ManifestBenchmark tests
asserting a three-line delegation to op.eval_roofline() that every real run
exercises, and the opt-in CUDA-events fallback test -- if that env var stopped
working the run fails loudly, and the test above it already pins that. What
stays is what would otherwise fail quietly: attribution keeping inter-kernel
gaps, failing closed on a dropped, missing, extra or reordered record, the
input pool handing out fresh pointers without corrupting values, and a genuine
kernel error reaching the caller instead of the fallback.
